refactor: dedupe _format_message_for_log into BrowserAgentMixin
navigator_n1.py, navigator_n1_custom_tools.py, and navigator_n1_memo.py
each defined a byte-for-byte identical _format_message_for_log method
(~17 lines) that PR #155 left out of the BrowserAgentMixin extraction
because navigator_n1_5.py's version differs stylistically. Since three
of the four copies are truly identical, move it into the mixin and let
navigator_n1_5.py's own differently-styled-but-equivalent
implementation continue to override it via normal MRO.
